requirements.txt
streamlit==1.38.0
pydantic==2.8.2
python-dotenv==1.0.1
pymupdf==1.24.10
pdfminer.six==20240706
rapidfuzz==3.9.6
loguru==0.7.2
black==24.8.0
ruff==0.6.9
pytest==8.3.2